# contracts/openapi/juno-native-v1.yaml
openapi: 3.1.0
info:
  title: Juno Native API
  version: 1.3.0
  description: >-
    Canonical web-to-native contract. Auth.js cookies are never native
    credentials.

    `version` is what installed clients compare against, so it moves only on a
    change that an older client cannot survive. Purely additive endpoints —
    /auth/password is one — leave it alone: a 1.3.0 client simply never calls
    them, whereas bumping it would fail that client's /auth/session check and
    sign the user out of an app that was working fine. The SHA-256 digest in the
    generated Swift contract still moves, so additive edits are not invisible.
servers:
  - url: /api/v1
security:
  - nativeBearer: []
paths:
...
  /library/attach:
    post:
      servers: [{ url: /api }]
      operationId: cloneNativeLibraryAttachments
      description: Clone owned attachment metadata for a future message without duplicating bytes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [attachmentIds]
              properties:
                attachmentIds:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items: { type: string }
      responses:
        '201':
          description: Fresh unlinked attachment rows in request order.
          content:
            application/json:
              schema:
                type: object
                required: [attachments]
                properties:
                  attachments: { type: array, items: { $ref: '#/components/schemas/LibraryAttachment' } }
        '400': { $ref: '#/components/responses/GeneralRouteError' }
        '401': { $ref: '#/components/responses/GeneralRouteError' }
        '404': { $ref: '#/components/responses/GeneralRouteError' }
...
  /agents/{agentId}/notes/{noteId}:
    parameters:
      - { in: path, name: agentId, required: true, schema: { type: string } }
      - { in: path, name: noteId, required: true, schema: { type: string } }
    patch:
      servers: [{ url: /api }]
      operationId: updateNativeAgentNote
      description: An edited note becomes the person's words (`source` = user).
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [content], properties: { content: { type: string, maxLength: 1000 } } }
      responses:
        '200':
          description: The note after the change.
          content: { application/json: { schema: { type: object, required: [note], properties: { note: { $ref: '#/components/schemas/AgentNote' } } } } }
        '400': { $ref: '#/components/responses/GeneralRouteError' }
        '401': { $ref: '#/components/responses/GeneralRouteError' }
        '404': { $ref: '#/components/responses/GeneralRouteError' }
    delete:
      servers: [{ url: /api }]
      operationId: deleteNativeAgentNote
      description: Gone from every later turn, which reads notes with `deletedAt` null.
      responses:
        '200': { description: Deleted. }
        '401': { $ref: '#/components/responses/GeneralRouteError' }
        '404': { $ref: '#/components/responses/GeneralRouteError' }
...
components:
  securitySchemes:
    nativeBearer: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
...
    GeneralRouteError:
      description: Legacy general-route error; normalization is tracked by GAP-006.
      content: { application/json: { schema: { $ref: '#/components/schemas/GeneralRouteErrorEnvelope' } } }
  schemas:
...
    LibraryAttachment:
      type: object
      additionalProperties: false
      required: [id, kind, fileName, mimeType, size, url, width, height]
      properties:
        id: { type: string }
        kind: { type: string, enum: [IMAGE, FILE] }
        fileName: { type: string }
        mimeType: { type: string }
        size: { type: integer, minimum: 0 }
        url: { type: string }
        width: { type: [integer, 'null'], minimum: 0 }
        height: { type: [integer, 'null'], minimum: 0 }
        createdAt: { type: string, format: date-time }
        conversationId: { type: [string, 'null'] }
...
    GeneralRouteErrorEnvelope:
      type: object
      additionalProperties: true
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        code: { type: string }
        retryable: { type: boolean }

    # -------------------------------------------------------------------------
    # Juno Work
    #
    # Field-for-field with the serializers in src/lib/work/serializers.ts and
    # src/lib/work/schedule.ts, which are the only functions allowed to build
    # these shapes. Two omissions there are load-bearing and are repeated here
    # so a client author reads them rather than discovering them: a run never
    # carries `claimedBy`/`claimedAt`/`leaseExpiresAt`/`idempotencyKey`, and a
    # file grant never carries `localPath`/`resolvedRealPath` on any surface
    # except the Mac that owns the path.
    #
    # `additionalProperties: false` throughout. These responses are built by
    # construction rather than spread from a row, so a column added next quarter
    # cannot ride out to a phone — and a contract that allowed extra keys would
    # quietly bless exactly that.
    # -------------------------------------------------------------------------
...
    AgentNote:
      type: object
      required: [id, agentId, content, source, createdAt, updatedAt]
      properties:
        id: { type: string }
        agentId: { type: string }
        content: { type: string }
        source: { type: string, enum: [user, agent, reflection] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...